+++ bidrag-transport/src/main/kotlin/no/nav/bidrag/transport/behandling/felles/grunnlag/ResultatFraVedtakGrunnlag.kt
@@ -1,5 +1,5 @@
 package no.nav.bidrag.transport.behandling.felles.grunnlag
 
 data class ResultatFraVedtakGrunnlag(
-    val vedtaksid: Long,
+    val vedtaksid: Int,
 ) : GrunnlagInnhold
